Refactored Map section
